+++ apps/DesktopStreamer/MainWindow.cpp
@@ -49,7 +49,9 @@
 #include <QScreen>
 
 #include <algorithm>
+#include <cmath>
 #include <iostream>
+#include <sstream>
 
 #ifdef _WIN32
 typedef __int32 int32_t;
@@ -69,12 +71,7 @@ typedef __int32 int32_t;
 
 namespace
 {
-const std::vector< std::pair< QString, QString > > defaultHosts = {
-    { "DisplayWall Ground floor", "bbpav02.epfl.ch" },
-    { "DisplayWall 3rd floor", "bbpav04.epfl.ch" },
-    { "DisplayWall 5th floor", "bbpav05.epfl.ch" },
-    { "DisplayWall 6th floor", "bbpav06.epfl.ch" }
-};
+const std::vector< std::pair< QString, QString > > defaultHosts = { _HOSTS };
 const QString streamButtonDefaultText = "Stream";
 const QString streamSelected = "Stream selected item(s)";
 }